Making the time entry collection a map that has the users as keys.  That will neatly coincide with our movement to splitting up the database by user when we write it to disk.
Starting to create some tests and code for writing our database to disk more intelligently.  Instead of one big file, it makes a lot of sense to break it into many small files.  A seemingly reasonable approach is 100 entries per file, up to 1000 files per directory, up to 1000 directories, and we split up the time entries by user.  This way, writes and reads are crazy fast and we can even review the database files easily.

Switching some of the last lingering longs to ints. (We don't need ints, we will do what it takes to avoid getting even close to the number 2 billion in this application) (Longs are twice the size of Ints, and we don't even need them, so toss 'em)

+++ src/main/kotlin/coverosR3z/timerecording/ITimeEntryPersistence.kt
@@ -10,7 +10,7 @@ interface ITimeEntryPersistence {
     /**
      * Provided a user and date, give the number of minutes they worked on that date
      */
-    fun queryMinutesRecorded(user: User, date: Date): Long
+    fun queryMinutesRecorded(user: User, date: Date): Int
     fun readTimeEntries(user: User): List<TimeEntry>
     fun readTimeEntriesOnDate(user: User, date: Date): List<TimeEntry>
